+++ lib/log_service.dart
@@ -1,9 +1,8 @@
 /// A Dart package for simplify the usage of the official `logging` package.
 library;
 
-export 'package:ansicolor/ansicolor.dart';
 export 'package:logging/logging.dart';
 
-export 'src/terminal.dart';
 export 'src/log_controller.dart';
-export 'src/log_service.dart';
+export 'src/logger_controller.dart';
+export 'src/logger_provider.dart';