+++ lib/src/log_controller.dart
@@ -1,6 +1,6 @@
 import 'package:logging/logging.dart';
 
-/// Settings of the logger instance.
+/// An interface for controlling logs.
 abstract class LogController {
   /// Effective level considering the levels established in this logger's
   /// parents (when [hierarchicalLoggingEnabled] is true).