+++ lib/src/terminal_stub.dart
@@ -1,6 +1,0 @@
-import 'terminal.dart';
-
-Terminal createTerminal() {
-  throw UnsupportedError(
-      "Can't create a terminal without dart:html or dart:io.");
-}