+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 1.0.0
+
+- Add `LogController` interface.
+- Add `LoggerController` mixin.
+- Add `LoggerProvider` mixin.
+
+## 1.0.0-dev.4
+
+- Move methods in `LogService` to `logger` filed.
+
+## 1.0.0-dev.3
+
+- Downgrade `meta` version to 1.9.1
+
+## 1.0.0-dev.2
+
+- Remove `Terminal` API.
+- Fix issues.
+
 ## 1.0.0-dev.1
 
 - Initial version.